+++ src/hooks/subscriptions/index.ts
@@ -10,3 +10,6 @@ export type { UseSubscribeReturn, SubscribeParams } from './useSubscribe';
 
 export { useManageSubscription } from './useManageSubscription';
 export type { UseManageSubscriptionReturn } from './useManageSubscription';
+
+export { useMySubscriptions } from './useMySubscriptions';
+export type { UseMySubscriptionsReturn } from './useMySubscriptions';